Added publisher for motor and servo

+++ micro_ros_raspberrypi_pico_sdk/hardware/Motor.cpp
@@ -52,8 +52,8 @@ void Motor::setSpeed(int speedPWM) {
     currentPwm = speedPWM;
 }
 
-double Motor::getSpeed() {
-    return static_cast<double>(currentPwm);//static_cast<double>(irSensor->getSpeed());//(2 * 3.14159 * 0.05 * (static_cast<double>(irSensor->getSpeed())/279));
+int Motor::getSpeed() {
+    return currentPwm;//static_cast<double>(irSensor->getSpeed());//(2 * 3.14159 * 0.05 * (static_cast<double>(irSensor->getSpeed())/279));
 }
 
 void Motor::updateDirection(bool inAValue, bool inBValue) {
